break pagination up into two modules
