+++ src/assets/drizzle/scripts/drizzle.js
@@ -1,7 +1,9 @@
 /* global document window */
+
 import 'prismjs';
 import './form-state-changer';
-import './pagination';
+import './pagination/long';
+import './pagination/default';
 
 const dom = {};
